webservices/serve_climate_news/requirements.txt
Flask==1.0.2
tensorflow==1.14.0
google-cloud-storage
gpt_2_simple
